Clean project structure (and more pair graph work)
Replace com.github.elliottuck.sbfst package with sbfst package. Also add
resources folder (under src/main) to store *.fst.txt files and other
resources so that the sbfst package only contains the source code. Move
class Test to the sbfst package in a subdirectory of src/test so that it
automatically runs with the other unit tests (also slightly modify this
class to integrate it with JUnit). Lastly, implement and test the
delta_i portion of the pair graph algorithm.