Add getM0 function
Add function to get component m0 for a given SCC as defined in Part IV
of Kim, McNaughton, McCloskey 1991 (and add tests for it). Note that the
tests are a little brittle, since equality is being tested on ArrayLists
and really we should be using Sets (this is a more general problem that
should be addressed in the future).