Fix bugs in isPiecewiseTestable()
Fix bugs related to checking the equality of states. Basically the issue
was that I was comparing states that belonged to different versions of a
graph (e.g. one graph might be the base graph and the other would be a
non-oriented copy of that base graph) by the equals() method. So even
though they are logically the same state, the comparison would fail. The
fix was to compare states by their integer ID, as this correctly
identifies the same states even if they belong to slightly different
graphs (i.e. graphs with altered edges).